SEARCH NEW algorithm + improved README.md
Added SEARCH NEW and SEARCH OLD. The SEARCH NEW algorithm is ≈85%
faster than the old one, so use it instead :) the only thing is that
the new algorithm always starts at 1 and there is nothing to do about
it but its worth it anyways. And a very improved README.md has been
added, explaining every function and its pros and cons.